Fixed the Camera not working correctly anymore after switching classes
The m_Targets variable in the CameraControl is now private and can be
manipulated with help of the public SetTarget-method.
The InitTargets-method is used to initialize the array.
Adjusted the Code in Player according to the new approach using the
SetTarget-method in CameraControl in order to replace the old reference.
Swapped the key-inputs used for changing classes.
Fixed a typo in MeleePlayer and made the Start-method in Player
protected in order to decrease the amount of duplicate code.
Also removed some unused imports.