The PlayerManager initializes/spawns the Player now, when  there is no Player attached to it as a child-object
Also adjusted the prefabs for the new context, finished the
"InputManager"-part of the PlayerManager for now and started to migrate
the class-change funtionality into the PlayerManager.
I also reforrmated some c#-classes to keep a better overview of the
changes that still need to be done in the process.