Completed the functionality of the PlayerManager for now
The inputs are set within the builtin Unity InputManager.
Setting up players into the scene has changed:
The only thing you have to do now is simply putting the Player1 and
Player2 prefab into the scene.
You can customize the first class to be used within the Unity inspector
by changing the order in the classes array.
The position of both players first spawns is now equal to the transform
of the Player1-/Player2-Gameobject.
The next step is improving the GameManager's Code, because its still
very clunky.