Bugfixing (stuck in animation, animation canceling & spawn attack bug)
It was possible to get stuck during the attacking animation of the
ranger. This was caused by a CD which was set too low. You also can now
longer cancel the warrior attack animation and still attack an enemy.
The colliders now get deactivted as soon as the attack animation isn't
played anymore. Fixed a bug where you can attack in all four directions
after spawning and not moving.