Improved the code quality of the PlayerManager
Exchanged the playerTransform for playerObject (GameObject), because it
fits better for most usages.
Improved the ChangeClass functionality with the static hasFairy Bool,
this Bool is now used to determine if one of the player is a fairy
instead of searching for the other players GameObject and check manually
if it has a Fairy script attached to it.
It seems like the bug with both palyers beaing able to switch to fairy
at the same time by pressing the button in the same frame is fixed now.
Now only one player switches to fairy, but his target is destroyed (it
is the old class instance of the other player before changing classes)
when this scenario occurs.