First and second level are now working again as intended + updated AI
Github/Unity merges binary files which causes broken scenes, please safe
everything you did not only in the scene but also in the prefab of the
current level. That is really important! Also if somebody merges scenes
or prefabs never let them automerge, you have to take one file. If you
don't do this we risk a huge brick like this time again.

AI code is now smoother and also works if the current enemy is disabled
at the start of a scene. Searching paths and players is now done with
try-methods. This now prevents some heavy bugs in level two, where the
spiders calculated paths before the players even spawned. It is
important to disable all trap enemies in the scene, not with code. If
you do it in C# there is already a path calculated before the enemy
entities are getting disabled.