Skeleton enemy now gets animated properly. Warrior now animated while using ability
Fixed a bug where the animator called for a path direction before a path
was generated by A*. Buildsettings are now restored to load the first
level. Warrior animator needs one more transition back from attacking to
idle with shield & shield should be in front of him to suggest he is
blocking now.