Cooldownmanager commented, AI bug fix
Most important functions in cooldown manager are now commented (also
added an explanation to a call to those functions in warrior). Fixed a
bug, which caused the AI to make too large steps when there is only a
short path remaining. Nerfed the fairy speedboost a little bit. Fixed
index out of bounds exception in subject, now all loops are working
correctly.