The UI is now fully working (still needs testing, there might be some bugs)
The core-component of the UI-functionality is the UIManager.
It contains all required UI-Elements that are dragged in from the scene
(player portrait, healthbar, etc.) and the sprites that are exchanged
with the ones contained in thos UI-Elements (WarriorPortrait-sprite for
example).
The "cooldown-animations" are implemented via IEnumerators and the
WaitForSeconds-method in order to be in sync with the CooldownManager
(-> show the cooldown most accuratly)
It also contains all methods, that are related to initializing and
updating the UI.
Those methods are getting called from the PlayerManager(-> Player
related stuff (most important the players current health/hitpoints)) and
the CooldownManager (-> Cooldowns).
The hitpoints of the MeleePlayer and RangedPlayer have been increased to
100 in order to fit the UI better.
Hitpoints seem to have no limit (can be negative). This should be fixed,
if it isn't already in another branch.
There is also a small thing I encountered regarding Skill Icons:
It seems, that there are no icons for the ranged attack (and the
elemental variations).
Apart from that case the elemental variations of the skills should also
work now (the Ranger has the MeleePlayers Attack Icon + variations atm,
they can be exchanged against more fitting ones by just dragging them in
the Unity Inspector).
The only thing remaing to be responsible are the shown controls for each
skill.
The "cooldown-animations" also need to be improved in terms of
smoothness, because they only update each second at the moment (and that
looks kinda awkward).