Added Controller-Support for the MainMenu
It's also possible to navigate through the MainMenu with the KeyBoard
now.
Only Player 1 can navigate through the menu.
You can change the inputs for the menu in the EventSystem-GameObject.
At the moment you navigate with the left stick or the w- and s-button.
Submit is set to the A/X (xbox/ps) and space.